Update parser to latest

+++ src/models/gamestate.rs
@@ -17,15 +17,15 @@ pub struct LivingCharacter {
 #[derive(Debug, Deserialize)]
 pub struct AliveData {
     #[serde(default, deserialize_with = "deserialize_eu4_float")]
-    pub gold: Option<f32>,
+    pub gold: Option<f64>,
     pub health: Option<f32>,
     pub income: Option<f32>,
 }
 
-pub(crate) fn deserialize_eu4_float<'de, D>(deserializer: D) -> Result<Option<f32>, D::Error>
+pub(crate) fn deserialize_eu4_float<'de, D>(deserializer: D) -> Result<Option<f64>, D::Error>
 where
     D: Deserializer<'de>,
 {
-    let val: Option<f32> = Option::deserialize(deserializer)?;
+    let val: Option<f64> = Option::deserialize(deserializer)?;
     val.map(reencode_float).map(Ok).transpose()
 }
